Responsive monkey patch.
- Removed leg.width and leg.zindex attributes
- Added leg.extraclass attributes
- Added default width and zindex in CSS
- Added media queries for leg in CSS
- Added check on leg's boundaries
- Make sure the arrows is always on target

This patch is a personal quick fix - do not merge with main branch as it is not compatible with the original author's view.

Also, it is only tested and working for orientation: bottom, align: center. All the other orientation I personally didn't need so I didn't bother (since it's not going to get merged anyway)